python/requirements.txt
psycopg2-binary==2.8.6 #ARM specific
pylint==2.15.5